% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_boem_wea_outlines.R
\docType{data}
\name{boem_wea_outlines}
\alias{boem_wea_outlines}
\title{BOEM Renewable Energy Lease Areas and Wind Planning Areas}
\format{
A \code{sf} collection with 55 features and 23 fields.
\describe{
\item{Geometry type}{MULTIPOLYGON}
\item{Dimension}{XY}
\item{Bounding box}{xmin: -158.6636 ymin: 20.7765 xmax: -70.0215 ymax: 43.9597}
\item{Geodetic CRS}{WGS 84}
\item{Source}{data-raw/boem-renewable-energy-geodatabase/BOEMWindLayers_4Download.gdb}
\item{Fields}{Field descriptions can be found \href{https://services1.arcgis.com/Hp6G80Pky0om7QvQ/ArcGIS/rest/services/BOEM_Wind_Planning_and_Lease_Areas/FeatureServer/layers}{here}.}
}
}
\source{
\url{https://www.boem.gov/renewable-energy/mapping-and-data/renewable-energy-gis-data}
}
\usage{
data('boem_wea_outlines')
}
\description{
A \code{sf} object containing the outlines for BOEM Renewable Energy Lease Areas (LEASE_STAGE = 'Active') and Wind Planning Areas (LEASE_STAGE = 'Planning').
}
\details{
BOEM wind lease area outlines were updated on 2022/6/30 (version 10.81) and BOEM wind planning area outlines were updated on 2022/7/26 (version 10.81).

There may be more up to date BOEM wind planning areas than those included in NEFSCspatial.
It is recommended that you reach out to the Wind Team at the NEFSC (\email{angela.silva@noaa.gov}) to confirm.
}
\keyword{datasets}
